requirements.txt
# Milvus Multimodal RAG Requirements

# Core dependencies
milvus>=2.3.0  # Milvus Lite (no Docker required)
pymilvus>=2.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# OpenAI for metadata extraction
openai>=1.0.0

# Image processing
Pillow>=9.5.0
requests>=2.31.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0

# Optional but recommended
# accelerate>=0.20.0  # For faster model loading
# bitsandbytes>=0.39.0  # For 8-bit quantization
